+++ contracts/interfaces/IPair.sol
@@ -25,8 +25,8 @@ interface IPair {
 
     struct State {
         Tokens reserves;
-        uint128 interest;
-        uint128 cdp;
+        uint112 interest;
+        uint112 cdp;
     }
 
     struct Pool {
@@ -141,8 +141,8 @@ interface IPair {
         uint256 maturity,
         address liquidityTo,
         address dueTo,
-        uint128 interestIncrease,
-        uint128 cdpIncrease
+        uint112 interestIncrease,
+        uint112 cdpIncrease
     )
         external
         returns (
@@ -162,8 +162,8 @@ interface IPair {
         uint256 maturity,
         address bondTo,
         address insuranceTo,
-        uint128 interestDecrease,
-        uint128 cdpDecrease
+        uint112 interestDecrease,
+        uint112 cdpDecrease
     ) external returns (Claims memory claimsOut);
 
     function withdraw(
@@ -178,8 +178,8 @@ interface IPair {
         address assetTo,
         address dueTo,
         uint128 assetOut,
-        uint128 interestIncrease,
-        uint128 cdpIncrease
+        uint112 interestIncrease,
+        uint112 cdpIncrease
     ) external returns (uint256 id, Due memory dueOut);
 
     function pay(
